+++ dartagnan/src/main/java/com/dat3m/dartagnan/witness/graphml/EdgeAttributes.java
@@ -1,4 +1,4 @@
-package com.dat3m.dartagnan.witness;
+package com.dat3m.dartagnan.witness.graphml;
 
 public enum EdgeAttributes {
