+++ dartagnan/src/main/java/com/dat3m/dartagnan/witness/graphviz/ExecutionGraphVisualizer.java
@@ -1,4 +1,4 @@
-package com.dat3m.dartagnan.utils.visualization;
+package com.dat3m.dartagnan.witness;
 
 import com.dat3m.dartagnan.expression.Expression;
 import com.dat3m.dartagnan.program.Register;
@@ -9,6 +9,8 @@ import com.dat3m.dartagnan.program.event.core.MemoryCoreEvent;
 import com.dat3m.dartagnan.program.event.metadata.MemoryOrder;
 import com.dat3m.dartagnan.verification.model.EventData;
 import com.dat3m.dartagnan.verification.model.ExecutionModel;
+import com.dat3m.dartagnan.witness.graphviz.Graphviz;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
