+++ svcomp/src/main/java/com/dat3m/svcomp/SVCOMPRunner.java
@@ -2,8 +2,8 @@ package com.dat3m.svcomp;
 
 import com.dat3m.dartagnan.parsers.witness.ParserWitness;
 import com.dat3m.dartagnan.utils.options.BaseOptions;
+import com.dat3m.dartagnan.witness.graphml.WitnessGraph;
 import com.dat3m.dartagnan.configuration.Property;
-import com.dat3m.dartagnan.witness.WitnessGraph;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.io.Files;
 import org.sosy_lab.common.configuration.Configuration;
@@ -25,7 +25,7 @@ import org.apache.logging.log4j.Logger;
 import static com.dat3m.dartagnan.configuration.OptionInfo.collectOptions;
 import static com.dat3m.dartagnan.configuration.OptionNames.*;
 import static com.dat3m.dartagnan.parsers.program.utils.Compilation.*;
-import static com.dat3m.dartagnan.witness.GraphAttributes.UNROLLBOUND;
+import static com.dat3m.dartagnan.witness.graphml.GraphAttributes.UNROLLBOUND;
 import static java.lang.Integer.parseInt;
 
 @Options
